refactor: Long이 아닌 Bigdecimal Type으로 변환하도록 수정

+++ src/main/java/kitchenpos/common/MoneyConverter.java
@@ -1,19 +1,20 @@
 package kitchenpos.common;
 
+import java.math.BigDecimal;
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 import kitchenpos.vo.Money;
 
 @Converter(autoApply = true)
-public class MoneyConverter implements AttributeConverter<Money, Long> {
+public class MoneyConverter implements AttributeConverter<Money, BigDecimal> {
 
     @Override
-    public Long convertToDatabaseColumn(Money money) {
-        return money.getAmount().longValue();
+    public BigDecimal convertToDatabaseColumn(Money money) {
+        return money.getAmount();
     }
 
     @Override
-    public Money convertToEntityAttribute(Long amount) {
+    public Money convertToEntityAttribute(BigDecimal amount) {
         return Money.valueOf(amount);
     }
 }
